renderer: Move shader loading to its own file
