+++ src/emulator/renderer/src/functions.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <crypto/hash.h>
 #include <glutil/object.h>
 
 #include <psp2/gxm.h>
@@ -14,6 +15,7 @@ struct MemState;
 namespace renderer {
     struct TextureCacheState;
 
+    typedef std::map<Sha256Hash, std::string> GLSLCache;
     typedef std::tuple<std::string, std::string> ProgramGLSLs;
     typedef std::map<ProgramGLSLs, SharedGLObject> ProgramCache;
 
@@ -28,6 +30,10 @@ namespace renderer {
     // Texture cache.
     bool init(TextureCacheState &cache);
     void cache_and_bind_texture(TextureCacheState &cache, const SceGxmTexture &gxm_texture, const MemState &mem, bool enabled);
+
+    // Shaders.
+    std::string load_fragment_shader(GLSLCache &cache, const SceGxmProgram &fragment_program, const char *base_path);
+    std::string load_vertex_shader(GLSLCache &cache, const SceGxmProgram &vertex_program, const char *base_path);
 
     // Texture formats.
     const GLenum *translate_swizzle(SceGxmTextureFormat fmt);
